requirements.txt
torch
torchvision
medmnist
matplotlib
seaborn
scikit-learn
tqdm
pandas
numpy
transformers
medmnist
faiss-cpu
Pillow
